Extract factory to separate artifact
