+++ auto-value-gson-factory/gradle.properties
@@ -1,0 +1,3 @@
+POM_ARTIFACT_ID=auto-value-gson-factory
+POM_NAME=AutoValue: Gson Extension Factory
+POM_PACKAGING=jar
